Colocación de los jugadores y opciones como iniciar reiniciar y atras creación de la clase hilos para el movimiento de los competidores y ahora empezar a crear el código de los reportes del juego 1

+++ src/Game_1/Swimmer.java
@@ -41,7 +41,7 @@ public class Swimmer {
         return time;
     }
 
-    public void setTime(long time) {
+    public void setTime(int time) {
         this.time = time;
     }
